Refactor: extract weather forecast item layouts to sunriseui module
